Create test utils. Test all toolbars.
