+++ tests/03_model_test.js
@@ -1,54 +1,19 @@
-
-function loadModel(client, fnames){
-
-    for (const name of fnames) {
-
-        console.log("Loading: " + name);
-        client.execute(
-            function (fname) {
-                _loadModel(fname);
-            }, [name], null
-        );
-
-        client.pause(1000);
-
-        client.getTitle(function(title) {
-            this.assert.ok(title.includes(name), "File: " + name + " is opened");
-        });
-    }
-
-}
+let test_utils = require('./test_utils');
 
 module.exports = {
 
     beforeEach : function (client) {
         client.url('http://localhost:8124/atompm').pause(300);
-
-
     },
 
     'Login' : function (client) {
-
-        client.execute(
-            function() {
-                UserManagement.login('testuser');
-            }, [], null
-        );
-
-        client.pause(300);
-
-        client.getTitle(function(title) {
-            this.assert.ok(title.includes("AToMPM - [Unnamed]"), "AToMPM is opened");
-        });
+        test_utils.login(client);
     },
 
     'Load model' : function (client) {
 
         let filename = 'Formalisms/ClassicDEVS/ClassicDEVS.model';
-
-
-        loadModel(client, [filename]);
-
+        test_utils.load_model(client, [filename]);
     },
 
     'Load two models' : function (client) {
@@ -58,9 +23,8 @@ module.exports = {
             'Formalisms/Annotation/AnnotationMM.model'
         ];
 
-        loadModel(client, filenames);
+        test_utils.load_model(client, filenames);
     },
-
 
     after : function (client) {
         client.end();
